+++ PowerEditor/src/ScintillaComponent/GoToLineDlg.cpp
@@ -100,13 +100,13 @@ INT_PTR CALLBACK GoToLineDlg::run_dlgProc(UINT message, WPARAM wParam, LPARAM)
 						}
 						else
 						{
-							int posToGoto = 0;
+							size_t posToGoto = 0;
 							if (line > 0)
 							{
 								// make sure not jumping into the middle of a multibyte character
 								// or into the middle of a CR/LF pair for Windows files
 								auto before = (*_ppEditView)->execute(SCI_POSITIONBEFORE, line);
-								posToGoto = static_cast<int>((*_ppEditView)->execute(SCI_POSITIONAFTER, before));
+								posToGoto = (*_ppEditView)->execute(SCI_POSITIONAFTER, before);
 							}
 							auto sci_line = (*_ppEditView)->execute(SCI_LINEFROMPOSITION, posToGoto);
 							(*_ppEditView)->execute(SCI_ENSUREVISIBLE, sci_line);
@@ -182,7 +182,7 @@ void GoToLineDlg::updateLinesNumbers() const
 	else
 	{
 		current = static_cast<unsigned int>((*_ppEditView)->execute(SCI_GETCURRENTPOS));
-		int currentDocLength = (*_ppEditView)->getCurrentDocLen();
+		size_t currentDocLength = (*_ppEditView)->getCurrentDocLen();
 		limit = static_cast<unsigned int>(currentDocLength > 0 ? currentDocLength - 1 : 0);
 	}
     ::SetDlgItemInt(_hSelf, ID_CURRLINE, current, FALSE);
